requirements.txt
streamlit
pandas
numpy
scikit-learn
plotly
nltk
python-dotenv
nomic
umap-learn